requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asgiref==1.1.2            # via channels, daphne
attrs==17.2.0             # via automat
autobahn==17.10.1         # via daphne
automat==0.6.0            # via twisted
certifi==2017.11.5        # via requests
channels==1.1.8
chardet==3.0.4            # via requests
constantly==15.1.0        # via twisted
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
coverage==4.4.2
daphne==1.3.0             # via channels
django-filter==1.1.0
django==1.11.7
djangorestframework==3.7.3
docutils==0.14            # via m2r
hyperlink==17.3.1         # via twisted
idna==2.6                 # via requests
incremental==17.5.0       # via twisted
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
m2r==0.1.12
markupsafe==1.0           # via jinja2
mistune==0.8.1            # via m2r
pytz==2017.3              # via django
pyyaml==3.12
requests==2.18.4
six==1.11.0               # via asgiref, autobahn, automat, txaio, unittest-xml-reporting
twisted==17.9.0           # via daphne
txaio==2.8.2              # via autobahn
unittest-xml-reporting==2.1.0
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
zope.interface==4.4.3     # via twisted
